Further reorg controls
